Refactor Site Editor code
