+++ packages/edit-site/src/components/block-editor/constants.js
@@ -1,0 +1,1 @@
+export const FOCUSABLE_ENTITIES = [ 'wp_template_part' ];
